Reorganized Event Listeners (task #4340)
